@@ -3,8 +3,8 @@
     "type": "module",
     "dependencies": {
         "@aolyang/eslint-config": "workspace:*",
-        "@types/react": "^19.0.6",
-        "eslint": "^9.17.0",
-        "react": "^19.0.0"
+        "@types/react": "^19.2.2",
+        "eslint": "^9.38.0",
+        "react": "^19.2.0"
     }
 }
