@@ -2,7 +2,7 @@
     "name": "test-typescript",
     "type": "module",
     "dependencies": {
-        "eslint": "^9.17.0",
-        "@aolyang/eslint-config": "workspace:*"
+        "@aolyang/eslint-config": "workspace:*",
+        "eslint": "^9.38.0"
     }
 }
