@@ -49,19 +49,19 @@
         "dev": "pnpm build --watch"
     },
     "dependencies": {
-        "@eslint-react/eslint-plugin": "^1.23.2",
-        "@stylistic/eslint-plugin": "^2.12.1",
-        "@typescript-eslint/eslint-plugin": "^8.19.0",
-        "@typescript-eslint/parser": "^8.19.0",
-        "astro-eslint-parser": "^1.1.0",
-        "eslint": "^9.17.0",
+        "@eslint-react/eslint-plugin": "^1.53.1",
+        "@stylistic/eslint-plugin": "^5.5.0",
+        "@typescript-eslint/eslint-plugin": "^8.46.2",
+        "@typescript-eslint/parser": "^8.46.2",
+        "astro-eslint-parser": "^1.2.2",
+        "eslint": "^9.38.0",
         "eslint-merge-processors": "^0.1.0",
         "eslint-plugin-astro": "^1.3.1",
-        "eslint-plugin-react-hooks": "^5.1.0",
-        "eslint-plugin-react-refresh": "^0.4.16",
+        "eslint-plugin-react-hooks": "^5.2.0",
+        "eslint-plugin-react-refresh": "^0.4.24",
         "eslint-plugin-simple-import-sort": "^12.1.1",
         "eslint-plugin-svelte": "^2.46.1",
-        "eslint-plugin-vue": "^9.32.0",
+        "eslint-plugin-vue": "^9.33.0",
         "eslint-processor-vue-blocks": "^0.1.2",
         "svelte-eslint-parser": "^0.43.0",
         "vue-eslint-parser": "^9.4.3"
