@@ -11,13 +11,13 @@
     },
     "devDependencies": {
         "@aolyang/eslint-config": "workspace:*",
-        "@types/node": "^22.10.5",
-        "eslint": "^9.17.0",
+        "@types/node": "^22.18.12",
+        "eslint": "^9.38.0",
         "eslint-typegen": "^0.3.2",
-        "globals": "^15.14.0",
-        "semver": "^7.6.3",
-        "tsup": "^8.3.5",
-        "tsx": "^4.19.2",
-        "zx": "^8.3.0"
+        "globals": "^15.15.0",
+        "semver": "^7.7.3",
+        "tsup": "^8.5.0",
+        "tsx": "^4.20.6",
+        "zx": "^8.8.5"
     }
 }
